Detach new entries creation from Render thread
Prevents the same hang-up from loading an existing library for when a
new library is being created or additional entries are being added. As
`add_new_files_runnable` is no longer blocking, an additional signal was
needed to be added for the program to know when it was safe to display
the library.