Update requirements.txt
Changed the version to 6.7.0 because it is marked as a fixed version for the but where it says:
```No QtMultimedia backends found. Only QMediaDevices, QAudioDevice, QSoundEffect, QAudioSink, and QAudioSource are available.
Failed to create QVideoSink "Not available"
Failed to initialize QMediaPlayer "Not available"
```
Bug Report is attached below and the change log from 6.7.0 -> 6.7.1 is listed. two CVE's were listed but they are not directly used by QT and only used if the developer is using them. I did not check to see if the code contains, but since you are trying to use sqlite DB and are not networking this should be in the clear.

https://bugreports.qt.io/browse/PYSIDE-2656
https://code.qt.io/cgit/qt/qtreleasenotes.git/about/qt/6.7.1/release-note.md
https://forum.qt.io/topic/157026/no-qtmultimedia-backends-found-on-macos-with-pyside-6-7-1